Add an haproxy sidecar configuration
This config relies on an environment variable PORT which can be one of
the following depending on the service :

   - 2001 : time
   - 2002 : user
   - 2003 : log
   - 2004 : front

The local service is expected to run at 10000 + this port. Outgoing
connections are processed on 2000 + this port. The stats are available
on port 30000 + this port and on the CLI in /tmp/sock-$PORT.